功能完善: - 删除部门时：如果部门下有员工，则不允许删除该部门，并给前端提示错误信息 练习完成

+++ src/main/java/org/example/service/impl/DeptServiceImpl.java
@@ -1,10 +1,13 @@
 package org.example.service.impl;
 
+import org.example.exception.BusinessException;
 import org.example.mapper.DeptMapper;
+import org.example.mapper.EmpMapper;
 import org.example.pojo.Dept;
 import org.example.service.DeptService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDateTime;
 import java.util.List;
@@ -14,14 +17,21 @@ public class DeptServiceImpl implements DeptService {
 
     @Autowired
     private DeptMapper deptMapper;
+    @Autowired
+    private EmpMapper empMapper;
 
     @Override
     public List<Dept> findAll() {
         return deptMapper.findAll();
     }
 
+    @Transactional
     @Override
     public void deleteById(Integer id) {
+        int count=empMapper.countEmpByDeptId(id);
+        if(count!=0){
+            throw new BusinessException("部门下有员工，不能删除");
+        }
         deptMapper.deleteById(id);
     }
